Time Bombs fix, Azazel's Lost Horn WIP
Forgot to add a check for if the player actually had Time Bombs. It now
properly checks for that.

A very WIP but working version of Azazel's Lost Horn, it's awesome!
Needs a lot of animation work and some fixes, but overall it works well.

Some minor file name changes and addition to items.xml, itempools.xml
and entities2.xml.